hot fix

+++ 0x0E-structures_typedef/dog.h
@@ -11,7 +11,7 @@
 struct dog
 {
 	char *name;
-	double *age;
+	float *age;
 	char *owner;
 };
 /**
